modif composants mediaCard et missions, changement de nom de simpleAppBar et MediaCard
