+++ src/App.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import './App.css';
 import Home from './Home';
 
 
@@ -11,8 +10,7 @@ class App extends Component {
     return (
       <div className="App">
         <Home/> 
-        
-        </div>
+      </div>
       
     );
   }
