+++ src/TrashList.js
@@ -1,0 +1,24 @@
+export const TrashList = [
+    {
+        title: "Formations secteur public",
+        text1: "lorem ipsum...",
+        text2: "lorem ipsum...",
+        text3: "lorem ipsum...",
+        img: ""
+    },
+    {
+        title: "Formations secteur privé",
+        text1: "lorem ipsum...",
+        text2: "lorem ipsum...",
+        text3: "lorem ipsum...",
+        img: ""
+    },
+    {
+        title: "Conseils, études et audits",
+        text1: "lorem ipsum...",
+        text2: "lorem ipsum...",
+        text3: "lorem ipsum...",
+        img: ""
+    }
+
+]
